prefer .js for correct mime type
I have not tracked down exactly why this trick works... for example I
don't know if it is something in the browser or something in deno.
Skimming the source code for deno's http server.ts I don't find
anything obvious sending mime types in the headers. But firefox
treated the files differently in local development mode when they
arrived with .js extensions.